Move open editors check from ScalaProject to ScalaEditor
So that ui-related operations are removed from ScalaProject API.
