+++ org.scala-ide.sdt.core/src/org/scalaide/ui/internal/editor/ScalaEditor.scala
@@ -1,10 +1,16 @@
 package org.scalaide.ui.internal.editor
 
-import org.eclipse.jface.text.rules.FastPartitioner
-import org.eclipse.jface.text.IDocumentPartitioner
-import org.eclipse.jdt.ui.text.IJavaPartitions
 import scala.tools.eclipse.contribution.weaving.jdt.ui.javaeditor.IScalaEditor
-import org.scalaide.core.internal.lexical._
+
+import org.eclipse.jdt.ui.text.IJavaPartitions
+import org.eclipse.ui.IEditorReference
+import org.eclipse.ui.IFileEditorInput
+import org.eclipse.ui.IWorkbenchPage
+import org.scalaide.core.IScalaPlugin
+import org.scalaide.core.IScalaProject
+import org.scalaide.core.internal.lexical.ScalaDocumentPartitioner
+import org.scalaide.core.internal.lexical.ScalaPartitions
+import org.scalaide.util.internal.Utils.WithAsInstanceOfOpt
 
 trait ScalaEditor extends IScalaEditor with ISourceViewerEditor with InteractiveCompilationUnitEditor {
 
@@ -22,4 +28,31 @@ object ScalaEditor {
     IJavaPartitions.JAVA_CHARACTER,
     ScalaPartitions.SCALA_MULTI_LINE_STRING)
 
+  import org.scalaide.util.internal.Utils.WithAsInstanceOfOpt
+
+  /**
+   * Checks whether there's at least one open editor related to given project
+   */
+  def projectHasOpenEditors(project: IScalaProject): Boolean = {
+    def hasOpenEditorForThisProject(page: IWorkbenchPage) = {
+      val editorRefs = page.getEditorReferences
+      editorRefs exists hasEqualProject
+    }
+
+    def hasEqualProject(editorRef: IEditorReference) = {
+      val isEqual = for {
+        editor <- Option(editorRef.getEditor( /*restore =*/ false))
+        input <- editor.getEditorInput.asInstanceOfOpt[IFileEditorInput]
+      } yield {
+        val file = input.getFile
+        project.underlying equals file.getProject
+      }
+      isEqual.getOrElse(false)
+    }
+
+    IScalaPlugin().getWorkbench().getWorkbenchWindows().exists { workbenchWindow =>
+      val pages = workbenchWindow.getPages()
+      pages exists hasOpenEditorForThisProject
+    }
+  }
 }
